move LB update to route locator
