+++ gateway-service/src/main/java/org/zowe/apiml/gateway/ribbon/GatewayRibbonConfig.java
@@ -21,6 +21,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
 import org.zowe.apiml.gateway.cache.ServiceCacheEvictor;
+import org.zowe.apiml.gateway.routing.ApimlRouteLocator;
 
 /**
  * Configuration of client side load balancing with Ribbon
@@ -60,13 +61,12 @@ public class GatewayRibbonConfig {
     @Autowired
     public ILoadBalancer ribbonLoadBalancer(IClientConfig config,
                                             ServerList<Server> serverList, ServerListFilter<Server> serverListFilter,
-                                            IRule rule, IPing ping, ServerListUpdater serverListUpdater,
-                                            ServiceCacheEvictor serviceCacheEvictor) {
+                                            IRule rule, IPing ping, ServerListUpdater serverListUpdater, ApimlRouteLocator apimlRouteLocator) {
         if (this.propertiesFactory.isSet(ILoadBalancer.class, ribbonClientName)) {
             return this.propertiesFactory.get(ILoadBalancer.class, config, ribbonClientName);
         }
         return new ApimlLoadBalancer<>(config, rule, ping, serverList,
-            serverListFilter, serverListUpdater, serviceCacheEvictor);
+            serverListFilter, serverListUpdater, apimlRouteLocator);
     }
 
 }
