+++ integration-tests/src/test/java/org/zowe/apiml/gatewayservice/AuthenticationOnDeploymentTest.java
@@ -219,23 +219,23 @@ public class AuthenticationOnDeploymentTest {
                         .then().statusCode(SC_OK);
                 });
                 System.out.println("3 OUT_OF_SERVICE");
-//                Thread.sleep(2000);
+                Thread.sleep(2000);
                 given().when()
                     .put("https://localhost:10011/eureka/apps/" + serviceId + "/" + host + ":" + serviceId + ":" + 5678 + "/status?value=UP")
                     .then().statusCode(SC_OK);
                 System.out.println("1 UP 2 OUT_OF_SERVICE");
-//                Thread.sleep(4000);
+                Thread.sleep(2000);
                 service1.getGatewayVerifyUrls().forEach(x ->
                     given()
                         .when().get(x + "/test")
-                        .then().statusCode(is(SC_OK))
+                        .then().body(is("")).statusCode(is(SC_OK))
                 );
 
 //                unregister service1
                 given().when().delete("https://localhost:10011/eureka/apps/" + serviceId + "/" + host + ":" + serviceId + ":" + 5678).then().statusCode(SC_OK);
 
                 System.out.println(" 2 OUT_OF_SERVICE");
-//                Thread.sleep(4000);
+                Thread.sleep(2000);
 //                set service2 UP
                 given().when()
                     .put("https://localhost:10011/eureka/apps/" + serviceId + "/" + host + ":" + serviceId + ":" + 5679 + "/status?value=UP")
@@ -243,7 +243,7 @@ public class AuthenticationOnDeploymentTest {
 
 //                call service2
                 System.out.println("1 UP 1 OUT_OF_SERVICE");
-//                Thread.sleep(4000);
+                Thread.sleep(2000);
                 service2.getGatewayVerifyUrls().forEach(x ->
                     given()
                         .when().get(x + "/test")
@@ -257,16 +257,16 @@ public class AuthenticationOnDeploymentTest {
 
 //                call service3
                 System.out.println("2 UP");
-//                Thread.sleep(4000);
+                Thread.sleep(2000);
                 service3.getGatewayVerifyUrls().forEach(x ->
                     given()
                         .when().get(x + "/test")
-                        .then().statusCode(is(SC_OK))
+                        .then().body(is("")).statusCode(is(SC_OK))
                 );
 
                 service1.start("UP");
                 System.out.println("3 UP");
-//                Thread.sleep(4000);
+                Thread.sleep(2000);
             }
 
         }
